#467 reorganize activities into categories
