+++ src/js/entities/activities/shared/Follow.js
@@ -1,4 +1,4 @@
-import Activity from './Activity';
+import Activity from '../Activity';
 import Move from './Move';
 
 /**
