+++ src/js/entities/activities/shared/Move.js
@@ -1,6 +1,6 @@
-import Activity from './Activity';
-import Util from '../../common/Util';
-import { TILE_WIDTH, TILE_HEIGHT } from '../../common/Const';
+import Activity from '../Activity';
+import Util from '../../../common/Util';
+import { TILE_WIDTH, TILE_HEIGHT } from '../../../common/Const';
 
 class Move extends Activity {
   /**
