+++ src/js/entities/activities/shared/RotateToTarget.js
@@ -1,4 +1,4 @@
-import Activity from './Activity';
+import Activity from '../Activity';
 
 /**
  * Constructor function to RotateToTarget
